Created bridge module and moved KotlinDslAdapter
